% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_udpipe.R
\name{annotation_udpipe}
\alias{annotation_udpipe}
\title{etiquetage UDPipe}
\usage{
annotation_udpipe(
  path,
  save_output = TRUE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{path}{string: Chemin du dossier contenant les differents corpus.}

\item{save_output}{boolean: Sauvegarde les resultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: ecrase et sauve de nouveaux les resultats}
}
\value{
DataFrame: corpus_annote avec les colonnes (mots || lemmes || POS || feats || Oeuvre)
}
\description{
etiquetage du corpus situe dans path
}
\examples{
corpus_annote <- annotation_udpipe(system.file("extdata", "curpus-test", package = "MotiveR"))

}
