% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_specificites.R
\name{calcul_specificites}
\alias{calcul_specificites}
\title{Calcul de spécificités}
\usage{
calcul_specificites(
  save_freq = TRUE,
  retrait_frequence_1 = TRUE,
  corpus_grams = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{save_freq}{boolean Suverage les résultats avec les fréquences}

\item{retrait_frequence_1}{boolean: Retire spécificité de fréquence 1 dans la sortie}

\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\description{
Calcul de spécificités
}
