% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choix_nb_ngrams.R
\name{choix_nb_ngrams}
\alias{choix_nb_ngrams}
\title{Choix ngrams}
\usage{
choix_nb_ngrams(
  n_grams,
  corpus = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{n_grams}{int choix de l'encodage en n-grams, entre 2 et 7}

\item{corpus}{data.frame corpus_motifs motifs pour chaque corpus}

\item{corpus_path}{string Chemin du csv contenant les corpus_motifs motifs pour chaque corpus}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\value{
DataFrame: corpus_annote avec les columns (mots || lemmes || POS || feats || Oeuvre)
}
\description{
Choix du nb de ngrams sur le corpus de motifs
}
\examples{
corpus_path <- system.file("extdata", "example_output", package = "MotiveR")
corpus_grams <- choix_nb_ngrams(2, corpus_path=corpus_path, save_output=FALSE)

}
