% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_histogram.R
\name{motifs_histogram}
\alias{motifs_histogram}
\title{Histogramme des motifs}
\usage{
motifs_histogram(
  corpus_grams = NULL,
  corpus_path = NULL,
  nmots = 25,
  freq = "rel"
)
}
\arguments{
\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}

\item{nmots}{int sélection du nombre de motifs à afficher}

\item{freq}{string "rel" pour fréquence relative ou "abs" pour fréquence absolue}
}
\description{
Fonction génération d'histogramme
}
