% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_nuage.R
\name{motifs_nuage}
\alias{motifs_nuage}
\title{Nuage des motifs}
\usage{
motifs_nuage(corpus_grams = NULL, corpus_path = NULL, nmots = 25, freq = "rel")
}
\arguments{
\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}

\item{nmots}{int sélection du nombre de motifs à afficher}

\item{freq}{string "rel" pour fréquence relative ou "abs" pour fréquence absolue}
}
\description{
Fonction génération de wordclouds
}
