% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_stats.R
\name{motifs_stats}
\alias{motifs_stats}
\title{Statistiques générales}
\usage{
motifs_stats(
  corpus_grams = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{corpus_path}{string Chemin du csv contenant les corpus_motifs motifs pour chaque corpus}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}

\item{corpus}{data.frame corpus_motifs motifs pour chaque corpus mots | motifs | Oeuvre}
}
\value{
DataFrame: Oeuvre | motifs | n (fréq absolue) | nb_total_mots (dans l'oeuvre) |
n_rel (fréquence relative) | spécificités oeuvre par oeuvre | pourcentage (présence du motif par rapport au reste du corpus)
}
\description{
Fonction génération de statistiques générales
}
