% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_tf_idf.R
\name{motifs_tf_idf}
\alias{motifs_tf_idf}
\title{TF-IDF plot}
\usage{
motifs_tf_idf(
  n_motifs = 20,
  plot_type = "group",
  corpus_grams = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{n_motifs}{int sélection du nombre de motifs à afficher}

\item{plot_type}{string "group" pour une visualisation groupée, "sep" pour une visualisation séparée}

\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}
}
\description{
Fonction génération TF-IDF
}
