% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MotiveR}
\alias{MotiveR}
\title{Identification of lexico-grammatical patterns}
\description{
Development of a textual analysis method based on the identification of lexico-grammatical patterns, known as the
"motifs" method, by Dominique Legallois (Université Sorbonne Nouvelle Paris III) and Antoine Silvestre de Sacy.
}
\author{
Dominique Legallois

Antoine de Sacy \email{antoinedesacy@gmail.com}

Bruno Spilak \email{bruno.spilak@gmail.com}
}
