% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_tf_idf.R
\name{plot_tf_idf}
\alias{plot_tf_idf}
\title{Visualisation des TF-IDF}
\usage{
plot_tf_idf(corpus_words_ngrams, n_motifs = 20, plot_type = "sep")
}
\arguments{
\item{corpus_words_ngrams}{data.frame sous format mots || motifs || Oeuvre}

\item{n_motifs}{int sélection du nombre de motifs à afficher}

\item{plot_type}{string "group" pour une visualisation groupée, "sep" pour une visualisation séparée (par défaut)}
}
\value{
un ggplot
}
\description{
Fonction de génération d'histogrammes TF-IDF
}
\examples{
corpus_words_ngrams = motifs_tf_idf(corpus_path="./output")
plot_tf_idf(corpus_words_ngrams, n_motifs = 20, plot_type = "sep")

}
