% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_acp.R
\name{prepare_acp}
\alias{prepare_acp}
\title{Prépare ACP}
\usage{
prepare_acp(
  freq_filter = 2,
  n_motifs = NULL,
  corpus_grams = NULL,
  corpus_path = NULL
)
}
\arguments{
\item{freq_filter}{int filtre de fréquence pour alléger les normalisations de fréquences. Les
motifs qui ont une fréquence au moins égale a freq_filter sont sélectionnés.}

\item{n_motifs}{int Nombre maximum de motifs par oeuvre pour réaliser l'ACP. Si spécifié,
freq_filter est ignoré.}

\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}
}
\value{
DataFrame: avec en colonne, Oeuvre1 | Oeuvre2 | ... | OeuvreN avec les motifs en ligne.
}
\description{
Prépare les données pour réaliser une ACP (ex: normalisation et filtrage).
}
\examples{
file_path <- system.file("extdata", "example_output", "corpus_motifs_grams.csv",
package="MotiveR")
corpus_grams <- data.table::fread(file_path, encoding = "UTF-8", header = TRUE)
data <- prepare_acp(freq_filter=2, corpus_grams=corpus_grams)

}
