% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_corpus_udpipe.R
\name{regex_corpus_udpipe}
\alias{regex_corpus_udpipe}
\title{Regex corpus UDPipe}
\usage{
regex_corpus_udpipe(
  corpus = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{corpus}{data.frame contenant les différents corpus.}

\item{corpus_path}{string Chemin du dossier contenant les différents corpus.}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\value{
DataFrame: corpus_motifs motifs pour chaque corpus
}
\description{
Transformation en motifs UDPpipe
}
