% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retour_texte_specificites_un_motif.R
\name{retour_texte_specificites_un_motif}
\alias{retour_texte_specificites_un_motif}
\title{Retour au texte d'un motif}
\usage{
retour_texte_specificites_un_motif(
  motif_cible,
  len_context,
  n_grams,
  corpus_grams = NULL,
  corpus_path = NULL,
  corpus_spec = NULL,
  corpus_spec_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{motif_cible}{string motif cible}

\item{len_context}{int nombre de mots du contexte a afficher}

\item{n_grams}{int n-grams}

\item{corpus_grams}{data.frame corpus_motifs motifs pour chaque corpus mots | motifs | Oeuvre}

\item{corpus_path}{string Chemin du csv contenant les corpus_motifs motifs pour chaque corpus}

\item{corpus_spec}{data.frame corpus specificites}

\item{corpus_spec_path}{string Chemin du csv contenant les specificites du corpus}

\item{save_output}{boolean: Sauvegarde les resultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: ecrase et sauve de nouveaux les resultats}
}
\value{
DataFrame: Oeuvre | motifs | n (freq absolue) | nb_total_mots (dans l'oeuvre) | n_rel (frequence relative) | specificites oeuvre par oeuvre | pourcentage (presence du motif par rapport au reste du corpus)
}
\description{
Fonction pour retour aux textes a partir d'un motif de la table de specificites
}
\examples{
corpus_annote <- retour_texte_specificites_un_motif(", a le NC", 4, 4)

}
