% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_motif_pipeline.R
\name{tag_motif_pipeline}
\alias{tag_motif_pipeline}
\title{Pipeline Motifs UDPipe}
\usage{
tag_motif_pipeline(path, corpus = NULL, save_output = TRUE, overwrite = FALSE)
}
\arguments{
\item{path}{string: Chemin du dossier contenant les différents corpus.}

\item{corpus}{string: Chemin du dossier contenant les différents corpus.}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\value{
DataFrame: corpus_annote avec les columns (mots || lemmes || POS || feats || Oeuvre)
}
\description{
Étiquetage du corpus et transformation en motifs UDPpipe
}
