% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_udpipe.R
\name{annotation_udpipe}
\alias{annotation_udpipe}
\title{Étiquetage UDPipe}
\usage{
annotation_udpipe(
  path,
  save_output = TRUE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{path}{string: Chemin du dossier contenant les différents corpus.}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\value{
DataFrame: corpus_annote avec les columns (mots || lemmes || POS || feats || Oeuvre)
}
\description{
Étiquetage du corpus situé dans path
}
