% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_acp.R
\name{motifs_acp}
\alias{motifs_acp}
\title{ACP des motifs}
\usage{
motifs_acp(
  plot_type = "var+motif",
  freq_filter = 2,
  n_motifs = NULL,
  n_obs = NULL,
  corpus_grams = NULL,
  corpus_path = NULL
)
}
\arguments{
\item{plot_type}{string: "var+motif" plot variables et motifs (défaut) "var" plot variables, "motif" plot motifs}

\item{freq_filter}{int filtre de fréquence pour alléger les normalisations de fréquences. Les motifs qui ont une fréquence au moins égale a freq_filter sont sélectionnés.}

\item{n_motifs}{int Nombre maximum de motifs par oeuvre pour réaliser l'ACP. Si spécifié, freq_filter est ignoré.}

\item{n_obs}{int Nombre maximum d'observations pour réaliser l'ACP. Si NULL, toutes les observations sont utilisées.}

\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}
}
\description{
Fonction d'analyse des motifs en composantes principales
}
