% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_tf_idf.R
\name{motifs_tf_idf}
\alias{motifs_tf_idf}
\title{TF-IDF transformation}
\usage{
motifs_tf_idf(
  corpus_grams = NULL,
  corpus_path = NULL,
  save_output = FALSE,
  save_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{corpus_grams}{data.frame sous format mots || motifs || Oeuvre}

\item{corpus_path}{string chemin du csv contenant les motifs en ngram}

\item{save_output}{boolean: Sauvegarde les résultats}

\item{save_path}{string: Chemin du fichier de sauvergarde}

\item{overwrite}{boolean: Écrase et sauve de nouveaux les résultats}
}
\value{
Dataframe avec colonnes Oeuvre || motifs || n || tf || idf || tf_idf
}
\description{
Fonction génération TF-IDF
}
\examples{
motifs_tf_idf(corpus_path="./output")

}
