% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MotiveR}
\alias{MotiveR}
\title{Identification de motifs lexico-grammaticaux}
\description{
Développement sous R d’une méthode d’analyse textuelle fondée sur l’identification de motifs
lexico-grammaticaux, méthode dite des « motifs » par Dominique Legallois
(Université Sorbonne Nouvelle Paris III) et Antoine de Sacy.
}
\author{
Dominique Legallois

Antoine de Sacy

Bruno Spilak \email{bruno.spilak@gmail.com}
}
